requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
absl-py==2.3.1
accelerate==1.1.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
asttokens==3.0.1
async-timeout==5.0.1
attrs==25.4.0
bitsandbytes==0.49.1
blobfile==3.1.0
certifi==2025.8.3
cfgv==3.5.0
charset-normalizer==3.4.3
clean-fid==0.1.35
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.3
datasets==4.5.0
diffusers==0.33.1
dill==0.4.0
distlib==0.4.0
ds==0.0.1
einops==0.8.1
exceptiongroup==1.3.1
executing==2.2.1
facexlib==0.3.0
filelock==3.20.3
filterpy==1.4.5
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2024.6.1
ftfy==6.3.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.76.0
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
icecream==2.1.10
identify==2.6.16
idna==3.10
imageio==2.37.0
imgaug==0.2.6
importlib_metadata==8.7.0
iniconfig==2.3.0
Jinja2==3.1.4
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
lightning-utilities==0.15.2
llvmlite==0.46.0
lmdb==1.7.3
lxml==6.0.1
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
networkx==3.3
ninja==1.13.0
nodeenv==1.10.0
numba==0.63.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai-clip==1.0.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.82
packaging==25.0
pandas==2.3.2
pillow==11.0.0
piq==0.7.0
platformdirs==4.4.0
pluggy==1.6.0
prdc==0.2
pre_commit==4.5.1
propcache==0.4.1
protobuf==6.33.2
psutil==7.0.0
pyarrow==23.0.0
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
ruff==0.15.0
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scikit-video==1.1.11
scipy==1.15.3
sentencepiece==0.2.1
sentry-sdk==2.37.1
six==1.17.0
smmap==5.0.2
sympy==1.13.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.24
tk==0.1.0
tokenizers==0.22.0
tomli==2.4.0
torch==2.4.1+cu121
torch-fidelity==0.3.0
torchaudio==2.4.1+cu121
torchmetrics==1.8.2
torchvision==0.19.1+cu121
tqdm==4.67.1
transformers==4.56.1
triton==3.0.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.36.1
wandb==0.21.3
wcwidth==0.2.13
Werkzeug==3.1.4
xxhash==3.6.0
yapf==0.43.0
yarl==1.22.0
zipp==3.23.0
